+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added `icon: Image` as a parameter to disclosue view button
 
 ## [1.13.0] - 2024-02-22Z
 - Updated `TextInputView` and `CurrencyInputView` by adding a hint label, updated fonts and colours